requirements.txt
paho-mqtt
requests
pyjwt
pyjwt[crypto]
cryptography
pycrypto
pillow
tinkerforge